ipcz: BufferPools and Fragments
Introduces BufferPool, an object which owns an extensible set of
DriverMemoryMapping objects, each identified by a unique BufferId within
the pool. In addition to exposing the mappings themselves, BufferPool
can also register BlockAllocators to help manage its mapped memory.

BufferPool exposes a simple API for dynamic allocation of different
sizes of blocks from across the pool, abstracting away the selection of
a specific buffer or BlockAllocator to use.

Allocated blocks are described by a FragmentDescriptor: a tuple of
(BufferId, offset, size). This is a suitable wire structure to convey
shared memory references across a NodeLink, where both sides will
maintain a BufferPool with the same set of BufferIds and mappings.

A Fragment is a FragmentDescriptor with a cached pointer to the mapped
memory it describes. FragmentDescriptors can be resolved to concrete
Fragments by querying the appropriate BufferPool.

Bug: 1299283
Change-Id: Ifc9c230d9a614f40abc9ad38c9180eef119afeb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3636566
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003796}
NOKEYCHECK=True
GitOrigin-RevId: a8964228e68fbcb0d29ff4062dfc1d71c850dee3