ipcz: Implement remote parcel queue accounting
For the common case of stable routes with a single central link, this
implements basic remote parcel queue accounting via RouterLink and
RouterLinkState.

This allows traps to monitor for consumption of parcels and parcel data
by a portal's remote peer, and it allows for applications to self-limit
outgoing transmissions in order to keep the remote queue within some
agreed-upon bounds.

Ihese features make it possible for Mojo data pipes to be reimplemented
on top of ipcz portals.

Bug: 1299283
Change-Id: I9e088f354d6bc9d616ba9b8c84946fb4c0dc8d88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3803785
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1033117}
NOKEYCHECK=True
GitOrigin-RevId: f03f83a73199622c1b6285c9291bbc31dffc6aa5