MojoIpcz: Tweak DataPipe peer closure behavior
Before this change, MojoIpcz emulated data pipe peer closure by having
it reflect whether the underlying control portal is dead or alive. This
is OK, but it's not strictly correct and it breaks some test
expectations when enabling MojoIpcz by default.

This CL corrects the behavior through a few combined changes:

- ipcz Routers no longer surface the peer closure status bit to
  the application until all expected inbound parcels have been
  received and queued in the router. This isn't strictly necessary,
  but it eliminates ambiguity in the event of some subtly incorrect
  Mojo data pipe API usage encountered in the wild.

- Data pipes no longer reflect portal live/dead status with their
  peer closure signal; instead they reflect the portal's own peer
  closure bit.

- Data pipes no longer serialize their peer closure state separately
  in the driver object header; instead it is derived from the portal
  status upon deserialization.

All of this is in preparation for enabling MojoIpcz by default in most
runtime environments and test suites on most platforms.

Bug: 1299283
Change-Id: I686a7bb24592aa4f240b5236e14031c40a7b3eda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4243001
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/main@{#1104822}
NOKEYCHECK=True
GitOrigin-RevId: ae9cdc10010bb22328956ed6c0087dab1ab0213b